Начало разделения на слои
- Отделение sign_request от параметров (обобщение над dict-ом в целом), теперь можно подписать любой запрос тем же алгоритмом

- Точка входа вынесена в cli.main, как первый внешний слой приложения в который поступают данные

- Большинство клиентских api функций были вынесены в music_api (этот слой будет переработан следующим шагом)

- Основная бизнес логика пока что так и остаётся в core.py
