+++ ymd/__main__.py
@@ -1,5 +1,5 @@
 import sys
 
-from .cli import main
+from .cli.main import main
 
 sys.exit(main())
