+++ ymd/cli/__init__.py
@@ -1,0 +1,3 @@
+'''
+    Фронтенд модуль для CLI интерфейса Yandex Music Downloader.
+'''
