+++ ymd/music_api/__init__.py
@@ -1,0 +1,6 @@
+'''
+    Модуль транспортного API слоя для работы с Yandex Music.
+
+    Содержит модели для сериализации и десериализации данных,
+    а также функции для получения информации о треках и их скачивания.
+'''
